requirements.txt
fastapi[standard]
pyTelegramBotAPI
contextlib2
pydantic
mysql-connector-python
pytest